+++ similarity/RunSimilarity.cpp
@@ -126,7 +126,7 @@ RunSimilarity::RunSimilarity(int argc, char *argv[]) {
   // Initialize the array of method names. We set it to 10 as max. We'll
   // most likely never have this many of similarity methods available.
   method = (char **) malloc(sizeof(char *) * 10);
-  min_sim = (double *) malloc(sizeof(double *) * 10);
+  min_sim = (float *) malloc(sizeof(float *) * 10);
 
   // The concatenated method.
   char * cmethod = NULL;
