+++ stats/swilk.h
@@ -1,7 +1,7 @@
 #ifndef _SWILK_
 #define _SWILK_
 
-#include <gsl/gsl_statistics.h>
+#include <gsl/gsl_statistics_float.h>
 #include <stdio.h>
 #include <string.h>
 #include <math.h>
@@ -10,7 +10,7 @@
 #include "../general/vector.h"
 #include "../general/error.h"
 
-double poly(const double *cc, int nord, double x);
-void swilk(double *vector, int n, double *w, double *pw, int *ifault);
+float poly(const float *cc, int nord, float x);
+void swilk(float *vector, int n, float *w, float *pw, int *ifault);
 
 #endif
